+++ rasa/nlu/featurizers/dense_featurizer/mitie_featurizer.py
@@ -3,7 +3,7 @@ import typing
 from typing import Any, List, Text
 
 from rasa.nlu.config import RasaNLUModelConfig
-from rasa.nlu.featurizers import Featurizer
+from rasa.nlu.featurizers.featurzier import Featurizer
 from rasa.nlu.tokenizers.tokenizer import Token
 from rasa.nlu.training_data import Message, TrainingData
 
@@ -11,20 +11,18 @@ if typing.TYPE_CHECKING:
     import mitie
 
 from rasa.nlu.constants import (
-    MESSAGE_RESPONSE_ATTRIBUTE,
-    MESSAGE_INTENT_ATTRIBUTE,
     MESSAGE_TEXT_ATTRIBUTE,
     MESSAGE_TOKENS_NAMES,
     MESSAGE_ATTRIBUTES,
-    MESSAGE_SPACY_FEATURES_NAMES,
-    MESSAGE_VECTOR_FEATURE_NAMES,
+    MESSAGE_VECTOR_DENSE_FEATURE_NAMES,
 )
 
 
 class MitieFeaturizer(Featurizer):
 
     provides = [
-        MESSAGE_VECTOR_FEATURE_NAMES[attribute] for attribute in MESSAGE_ATTRIBUTES
+        MESSAGE_VECTOR_DENSE_FEATURE_NAMES[attribute]
+        for attribute in MESSAGE_ATTRIBUTES
     ]
 
     requires = [MESSAGE_TOKENS_NAMES[attribute] for attribute in MESSAGE_ATTRIBUTES] + [
@@ -59,9 +57,11 @@ class MitieFeaturizer(Featurizer):
                         attribute_tokens, mitie_feature_extractor
                     )
                     example.set(
-                        MESSAGE_VECTOR_FEATURE_NAMES[attribute],
-                        self._combine_with_existing_features(
-                            example, features, MESSAGE_VECTOR_FEATURE_NAMES[attribute]
+                        MESSAGE_VECTOR_DENSE_FEATURE_NAMES[attribute],
+                        self._combine_with_existing_dense_features(
+                            example,
+                            features,
+                            MESSAGE_VECTOR_DENSE_FEATURE_NAMES[attribute],
                         ),
                     )
 
@@ -73,9 +73,11 @@ class MitieFeaturizer(Featurizer):
             mitie_feature_extractor,
         )
         message.set(
-            MESSAGE_VECTOR_FEATURE_NAMES[MESSAGE_TEXT_ATTRIBUTE],
-            self._combine_with_existing_features(
-                message, features, MESSAGE_VECTOR_FEATURE_NAMES[MESSAGE_TEXT_ATTRIBUTE]
+            MESSAGE_VECTOR_DENSE_FEATURE_NAMES[MESSAGE_TEXT_ATTRIBUTE],
+            self._combine_with_existing_dense_features(
+                message,
+                features,
+                MESSAGE_VECTOR_DENSE_FEATURE_NAMES[MESSAGE_TEXT_ATTRIBUTE],
             ),
         )
 
@@ -97,10 +99,7 @@ class MitieFeaturizer(Featurizer):
         feature_extractor: "mitie.total_word_feature_extractor",
     ) -> np.ndarray:
 
-        vec = np.zeros(self.ndim(feature_extractor))
+        vec = []
         for token in tokens:
-            vec += feature_extractor.get_feature_vector(token.text)
-        if tokens:
-            return vec / len(tokens)
-        else:
-            return vec
+            vec.append(feature_extractor.get_feature_vector(token.text))
+        return np.array(vec)