+++ tests/nlu/training/test_train.py
@@ -34,7 +34,6 @@ def pipelines_for_tests():
                 "SpacyTokenizer",
                 "MitieFeaturizer",
                 "SpacyFeaturizer",
-                "NGramFeaturizer",
                 "RegexFeaturizer",
                 "CountVectorsFeaturizer",
                 "MitieEntityExtractor",